chore: move the accept method into a createAttendanceValidator class
