+++ src/main/java/school/hei/haapi/service/scheduler/AttendanceScheduler.java
@@ -39,12 +39,6 @@ public class AttendanceScheduler {
   @Scheduled(fixedDelay = 60000)
   @Transactional
   public synchronized void checkAttendancesEachDays() {
-    System.out.println(
-        "scheduler running at "
-            + LocalDateTime.now().getHour()
-            + ":"
-            + LocalDateTime.now().getMinute()
-            + " ...");
     List<CourseSession> courseSessions = attendanceSchedulerUtils.findCourseSessionOfTheDay();
 
     courseSessions.forEach(
